refactor to look more like OMF
