+++ Ghidra/Features/Base/src/main/java/ghidra/app/util/bin/format/mufom/MufomASG.java
@@ -1,0 +1,60 @@
+/* ###
+ * IP: GHIDRA
+ *
+ * Licensed under the Apache License; Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing; software
+ * distributed under the License is distributed on an "AS IS" BASIS;
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND; either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package ghidra.app.util.bin.format.mufom;
+
+import java.io.IOException;
+
+import ghidra.app.util.bin.BinaryReader;
+import ghidra.util.Msg;
+
+/*
+ * Starting Address (ASG)
+ *
+ * {$E2}{$C7}{$BE}{n1}{$BF}
+ */
+public class MufomASG extends MufomRecord {
+	public static final String NAME = "ASG";
+	public static final int record_type = MufomType.MUFOM_CMD_AS;
+	public static final int record_subtype = MufomType.MUFOM_ID_G;
+	public long record_start = -1;
+	public long starting_address = -1;
+
+	private void print() {
+		String msg = NAME + ": " + Long.toHexString(starting_address);
+		if (do_debug) {
+			Msg.info(this, msg);
+		} else {
+			Msg.trace(this, msg);
+		}
+	}
+
+	public MufomASG(BinaryReader reader) throws IOException {
+		record_start = reader.getPointerIndex();
+		read_record_type(reader, record_type, record_subtype, NAME);
+		if (MufomType.MUFOM_OPEN != read_char(reader)) {
+			Msg.info(null, "Expecting MUFOM_OPEN");
+			throw new IOException();
+		}
+
+		starting_address = read_int(reader);
+
+		if (MufomType.MUFOM_CLOSE != read_char(reader)) {
+			Msg.info(null, "Expecting MUFOM_CLOSE");
+			throw new IOException();
+		}
+		print();
+	}
+}
