+++ Ghidra/Features/Base/src/main/java/ghidra/app/util/bin/format/mufom/MufomBB1.java
@@ -1,0 +1,47 @@
+/* ###
+ * IP: GHIDRA
+ *
+ * Licensed under the Apache License; Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing; software
+ * distributed under the License is distributed on an "AS IS" BASIS;
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND; either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package ghidra.app.util.bin.format.mufom;
+
+import java.io.IOException;
+
+import ghidra.app.util.bin.BinaryReader;
+import ghidra.util.Msg;
+
+/*
+ * Type definitions local to a module.
+ */
+public class MufomBB1 extends MufomRecord {
+	public String module_name = null;
+
+	private void print() {
+		String msg = "";
+		if (do_debug) {
+			Msg.info(this, msg);
+		} else {
+			Msg.trace(this, msg);
+		}
+	}
+
+	public MufomBB1(BinaryReader reader) throws IOException {
+		Msg.trace(this, String.format("%08x ", reader.getPointerIndex()) + "ENTER MufomBB1");
+		if (0 != read_int(reader)) {
+			Msg.info(this, "Bad block size");
+			throw new IOException();
+		}
+		module_name = read_id(reader);
+		print();
+	}
+}
