+++ Ghidra/Features/Base/src/main/java/ghidra/app/util/bin/format/mufom/MufomSB.java
@@ -1,0 +1,61 @@
+/* ###
+ * IP: GHIDRA
+ *
+ * Licensed under the Apache License; Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing; software
+ * distributed under the License is distributed on an "AS IS" BASIS;
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND; either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package ghidra.app.util.bin.format.mufom;
+
+import java.io.IOException;
+
+import ghidra.app.util.bin.BinaryReader;
+import ghidra.util.Msg;
+
+/*
+ * 10.1 SB (section begin) Command
+ *
+ * SB-command → “SB” section-number “.”
+ * section-number → hexnumber
+ *
+ * {$E5}{n1}
+ */
+public class MufomSB extends MufomRecord {
+	public static final String NAME = "SB";
+	public static final int record_type = MufomType.MUFOM_CMD_SB;
+	public static final int record_subtype = -1;
+	public long record_start = -1;
+	public long section_number = -1;
+
+	public static boolean check(BinaryReader reader) throws IOException {
+		long offset = reader.getPointerIndex();
+		if (record_type == reader.readUnsignedByte(offset + 0)) {
+			return true;
+		}
+		return false;
+	}
+
+	private void print() {
+		String msg = NAME + ": " + section_number;
+		if (do_debug) {
+			Msg.info(this, msg);
+		} else {
+			Msg.trace(this, msg);
+		}
+	}
+
+	public MufomSB(BinaryReader reader) throws IOException {
+		record_start = reader.getPointerIndex();
+		read_record_type(reader, record_type, record_subtype, NAME);
+		section_number = read_int(reader);
+		print();
+	}
+}
